allow non integer sample rates

+++ monitor/cli/gpu/memory.py
@@ -71,8 +71,8 @@ class GPUMemory(Application):
     ALLOW_NOARGS = True
     interface = Interface(PROGRAM, USAGE, HELP)
 
-    sample_rate: int = 1
-    interface.add_argument('-s', '--sample-rate', type=int, default=sample_rate)
+    sample_rate: float = 1
+    interface.add_argument('-s', '--sample-rate', type=float, default=sample_rate)
 
     format_plain: bool = True
     format_csv: bool = False
